+++ xlcalculator/xlfunctions/information.py
@@ -11,6 +11,25 @@ def ISBLANK(cell: func_xltypes.XlAnything) -> func_xltypes.Boolean:
 
 @xl.register()
 @xl.validate_args
+def ISEVEN(num: func_xltypes.XlNumber) -> func_xltypes.Boolean:
+    """Returns TRUE if number is even, or FALSE if number is odd.
+
+    https://support.microsoft.com/en-us/office/
+        iseven-function-aa15929a-d77b-4fbb-92f4-2f479af55356
+    """
+
+    if int(num) == 1:
+        return False
+
+    elif (int(num) % 2) == 0:
+        return True
+
+    else:
+        return False
+
+
+@xl.register()
+@xl.validate_args
 def ISTEXT(cell: func_xltypes.XlAnything) -> func_xltypes.Boolean:
     """Returns True if the cell is text.
     """
@@ -32,3 +51,22 @@ def ISNA(cell) -> func_xltypes.Boolean:
 
     """
     return isinstance(cell, xlerrors.NaExcelError)
+
+
+@xl.register()
+@xl.validate_args
+def ISODD(num: func_xltypes.XlNumber) -> func_xltypes.Boolean:
+    """Returns TRUE if number is odd, or FALSE if number is even.
+
+    https://support.microsoft.com/en-us/office/
+        isodd-function-1208a56d-4f10-4f44-a5fc-648cafd6c07a
+    """
+
+    if int(num) == 1:
+        return True
+
+    elif (int(num) % 2) == 0:
+        return False
+
+    else:
+        return True