+++ CHANGES.rst
@@ -6,7 +6,7 @@ CHANGES
 0.2.12 (2020-11-23)
 -------------------
 
-- Add functions: ``PV()``, ``XIRR``
+- Add functions: ``PV()``, ``XIRR``, ``ISEVEN()``, ``ISODD()``
 
 
 0.2.11 (2020-11-16)